💾 Feat(Navigation): Pin scaffold from multi-pages
