+++ KitX Website Flutter/kitx_website/lib/pages/milestone_page.dart
@@ -1,8 +1,4 @@
-import 'dart:html';
-
 import 'package:flutter/material.dart';
-import 'package:get/get.dart';
-import 'package:kitx_website/pages/controls/drawer.dart';
 
 class MilestonesPage extends StatefulWidget {
   const MilestonesPage({super.key});
@@ -19,14 +15,6 @@ class _MilestonesPageState extends State<MilestonesPage> {
 
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: Text('Milestones_Page'.tr),
-      ),
-      drawer: AppDrawer(),
-      drawerEnableOpenDragGesture: true,
-      drawerEdgeDragWidth: (window.innerWidth ?? 300) / 2 + 50,
-      body: Center(child: Text('Under developing ...')),
-    );
+    return Center(child: Text('Under developing ...'));
   }
 }
