+++ KitX Website Flutter/kitx_website/lib/pages/navigation_page.dart
@@ -1,6 +1,10 @@
+import 'dart:html';
+
 import 'package:community_material_icon/community_material_icon.dart';
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
+import 'package:kitx_website/pages/controls/drawer.dart';
+import 'package:kitx_website/pages/controls/navigation_page_actions.dart';
 import 'package:kitx_website/pages/home_page.dart';
 import 'package:kitx_website/pages/milestone_page.dart';
 import 'package:kitx_website/pages/plugins_page.dart';
@@ -18,14 +22,30 @@ class _NavigationPageState extends State<NavigationPage> {
   double _buttonsPadding = 10;
   var _pageController = PageController(initialPage: 0);
 
+  void checkWindowWidth() {
+    var width = window.innerWidth;
+    if (width != null) {
+      if (width <= 500 && showTextOfLinkButton.value) {
+        showTextOfLinkButton.value = false;
+      } else if (width > 500 && !showTextOfLinkButton.value) {
+        showTextOfLinkButton.value = true;
+      }
+    }
+  }
+
   @override
   void initState() {
     super.initState();
     app.navPageController = _pageController;
+
+    window.onResize.listen((event) => checkWindowWidth());
+
+    checkWindowWidth();
   }
 
   @override
   Widget build(BuildContext context) {
+    themeButtonInLightMode.value = !context.isDarkMode;
     return Scaffold(
       body: Row(
         children: <Widget>[
@@ -81,16 +101,44 @@ class _NavigationPageState extends State<NavigationPage> {
           const VerticalDivider(thickness: 1, width: 1),
           Expanded(
             child: PageView(
-              children: const [
-                const ClipRect(child: const HomePage()),
-                const ClipRect(child: const PluginsPage()),
-                // const ClipRect(child: const DocsPage()),
-                const ClipRect(child: const MilestonesPage()),
-                const ClipRect(child: const SettingsPage()),
+              physics: NeverScrollableScrollPhysics(),
+              children: [
+                Scaffold(
+                  appBar: AppBar(
+                    title: Obx(
+                      () => Text(
+                        showTextOfLinkButton.value ? 'KitX Website' : 'KitX',
+                        overflow: showTextOfLinkButton.value ? TextOverflow.fade : TextOverflow.ellipsis,
+                      ),
+                    ),
+                    actions: getActions(context),
+                    leading: Builder(
+                      builder: (BuildContext context) {
+                        return IconButton(
+                          icon: const Icon(Icons.menu),
+                          onPressed: Scaffold.of(context).openDrawer,
+                          //   tooltip: '',
+                        );
+                      },
+                    ),
+                  ),
+                  drawer: AppDrawer(),
+                  drawerEnableOpenDragGesture: true,
+                  drawerEdgeDragWidth: (window.innerWidth ?? 300) / 2 + 50,
+                  body: PageView(
+                    children: const [
+                      const ClipRect(child: const HomePage()),
+                      const ClipRect(child: const PluginsPage()),
+                      // const ClipRect(child: const DocsPage()),
+                      const ClipRect(child: const MilestonesPage()),
+                      const ClipRect(child: const SettingsPage()),
+                    ],
+                    controller: _pageController,
+                    physics: const NeverScrollableScrollPhysics(),
+                    scrollDirection: Axis.vertical,
+                  ),
+                ),
               ],
-              controller: _pageController,
-              physics: const NeverScrollableScrollPhysics(),
-              scrollDirection: Axis.vertical,
             ),
           ),
         ],
