+++ KitX Website Flutter/kitx_website/lib/pages/plugins_page.dart
@@ -1,8 +1,4 @@
-import 'dart:html';
-
 import 'package:flutter/material.dart';
-import 'package:get/get.dart';
-import 'package:kitx_website/pages/controls/drawer.dart';
 
 class PluginsPage extends StatefulWidget {
   const PluginsPage({super.key});
@@ -19,14 +15,6 @@ class _PluginsPageState extends State<PluginsPage> {
 
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: Text('Plugins_Page'.tr),
-      ),
-      drawer: AppDrawer(),
-      drawerEnableOpenDragGesture: true,
-      drawerEdgeDragWidth: (window.innerWidth ?? 300) / 2 + 50,
-      body: Center(child: Text('Under developing ...')),
-    );
+    return Center(child: Text('Under developing ...'));
   }
 }
